feat(wire): add sqlx feature for optional sqlx dep
wire's sqlx dependency and FromRow derives are now optional and guarded
by the new sqlx feature, which is passed through to lib. cli and ui do
not set this feature but api does.

This is necessary because ui is a dioxus app and supposed to run on the
web platform but sqlx seems to be incompatible with wasm. It also avoids
unnecessary compilation of sqlx for client side users like the cli.

Signed-off-by: Sandro-Alessio Gierens <sandro@gierens.de>